Ok, recording and playback of velocity is working with melodizers and external instrument input.
Haven't tested the extended OSC message yet, but should work too.

Next is aftertouch message...

Pretty cool to record velocity sensitive keyboard into melodizer WITH velocity info.

+++ src/jklabs/monomic/MonomeListener.java
@@ -62,7 +62,7 @@ public interface MonomeListener {
 	 * @param y
 	 * @param velocity
 	 */
-	public void monomeXPressed(int x, int y, float velocity);
+	public void monomeXPressed(int x, int y, int velocity);
 	
 	/**
 	 * Implement in your listener class
