+++ src/mtn/sevenuplive/modes/events/KeyTransposeGroupEvent.java
@@ -29,7 +29,7 @@ public class KeyTransposeGroupEvent implements Event {
 	private int group;
 	private int key_x;
 	private int key_y;
-	private float velocity;
+	private int velocity;
 	
 	public int getGroup() {
 		return group;
@@ -43,13 +43,13 @@ public class KeyTransposeGroupEvent implements Event {
 		return key_y;
 	}
 	
-	public float getVelocity() {
+	public int getVelocity() {
 		return velocity;
 	}
 	
 	public KeyTransposeGroupEvent() {}
 		
-	public KeyTransposeGroupEvent(int group, int key_x, int key_y, float velocity) {
+	public KeyTransposeGroupEvent(int group, int key_x, int key_y, int velocity) {
 		this.group = group;
 		this.key_x = key_x;
 		this.key_y = key_y;
